add: synk security fixes

+++ includes/class-ur-preview.php
@@ -221,7 +221,7 @@ class UR_Preview {
 			$html  = '';
 			$html .= '<div class="ur-preview-content">';
 			$html .= '<span class="ur-form-preview-title">';
-			$html .= esc_html(get_the_title( $form_id ));
+			$html .= esc_html( get_the_title( $form_id ) );
 			$html .= '</span>';
 
 			if ( function_exists( 'apply_shortcodes' ) ) {
@@ -314,17 +314,18 @@ class UR_Preview {
 		$option_name    = isset( $_GET['ur_email_preview'] ) ? sanitize_text_field( wp_unslash( $_GET['ur_email_preview'] ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 		$email_template = isset( $_GET['ur_email_template'] ) ? sanitize_text_field( wp_unslash( $_GET['ur_email_template'] ) ) : ''; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
 
-		$class_name = 'UR_Settings_' . str_replace( ' ', '_', ucwords( str_replace( '_', ' ', $option_name ) ) );
 		/**
 		 * Applies a filter to modify the email classes.
 		 */
 		$emails = apply_filters( 'user_registration_email_classes', array() );
 
+		$class_name = 'UR_Settings_' . str_replace( ' ', '_', ucwords( str_replace( '_', ' ', $option_name ) ) );
+
 		if ( isset( $emails[ $class_name ] ) && ! class_exists( $class_name ) ) {
 			$class_name = get_class( $emails[ $class_name ] );
 		}
 
-		if ( ! class_exists( $class_name ) ) {
+		if ( ! class_exists( $class_name ) && isset( $emails[ $class_name ] ) ) {
 			echo '<h3>' . esc_html__( 'Something went wrong. Please verify if the email you want to preview exists or addon it is associated with is activated.', 'user-registration' ) . '</h3>';
 		} else {
 			$class_instance = new $class_name();
@@ -347,12 +348,20 @@ class UR_Preview {
 			 * @param string $email_content The email message content.
 			 */
 			$email_content = apply_filters( 'user_registration_process_smart_tags', $email_content );
-
+			$allowed       = array_merge(
+				wp_kses_allowed_html( 'post' ),
+				array(
+					'style' => array(),
+					'head'  => array(),
+					'html'  => array(),
+					'body'  => array(),
+				)
+			);
 			ur_get_template(
 				'email-preview.php',
 				array(
-					'email_content'  => $email_content,
-					'email_template' => $email_template,
+					'email_content'  => wp_kses( $email_content, $allowed ),
+					'email_template' => sanitize_text_field( $email_template ),
 				)
 			);
 		}
